+++ src/main/java/br/com/zalf/prolog/webservice/frota/checklist/ChecklistDaoImpl.java
@@ -8,9 +8,7 @@ import br.com.zalf.prolog.webservice.database.DatabaseConnection;
 import br.com.zalf.prolog.webservice.frota.checklist.OLD.AlternativaChecklist;
 import br.com.zalf.prolog.webservice.frota.checklist.OLD.ModeloChecklist;
 import br.com.zalf.prolog.webservice.frota.checklist.OLD.PerguntaRespostaChecklist;
-import br.com.zalf.prolog.webservice.frota.checklist.model.AlternativaChecklistStatus;
-import br.com.zalf.prolog.webservice.frota.checklist.model.Checklist;
-import br.com.zalf.prolog.webservice.frota.checklist.model.NovoChecklistHolder;
+import br.com.zalf.prolog.webservice.frota.checklist.model.*;
 import br.com.zalf.prolog.webservice.frota.checklist.model.farol.DeprecatedFarolChecklist;
 import br.com.zalf.prolog.webservice.frota.checklist.modelo.ChecklistModeloDao;
 import br.com.zalf.prolog.webservice.frota.veiculo.VeiculoDao;
@@ -27,7 +25,7 @@ import java.util.*;
 
 import static br.com.zalf.prolog.webservice.commons.util.StatementUtils.bindValueOrNull;
 
-public class ChecklistDaoImpl extends DatabaseConnection implements ChecklistDao {
+public final class ChecklistDaoImpl extends DatabaseConnection implements ChecklistDao {
 
     public ChecklistDaoImpl() {
 
@@ -35,43 +33,14 @@ public class ChecklistDaoImpl extends DatabaseConnection implements ChecklistDao
 
     @NotNull
     @Override
-    public Long insert(@NotNull final Connection conn,
-                       @NotNull final Checklist checklist,
-                       final boolean deveAbrirOs) throws Throwable {
-        return internalInsertChecklist(conn, checklist, deveAbrirOs);
-    }
-
-    @NotNull
-    @Override
-    public Long insert(Checklist checklist) throws SQLException {
+    public Long insert(@NotNull Checklist checklist) throws SQLException {
         Connection conn = null;
-        try {
-            conn = getConnection();
-            conn.setAutoCommit(false);
-            final Long codChecklist = internalInsertChecklist(conn, checklist, true);
-            conn.commit();
-            return codChecklist;
-        } catch (final Throwable t) {
-            if (conn != null) {
-                conn.rollback();
-            }
-
-            // Como esse método ainda não está refatorado para retornar um Throwable, encapsulamos o retorno em uma
-            // SQLException.
-            throw new SQLException(t);
-        } finally {
-            close(conn);
-        }
-    }
-
-    @NotNull
-    private Long internalInsertChecklist(@NotNull final Connection conn,
-                                         @NotNull final Checklist checklist,
-                                         final boolean deveAbrirOs) throws Throwable {
         PreparedStatement stmt = null;
         ResultSet rSet = null;
         final VeiculoDao veiculoDao = Injection.provideVeiculoDao();
         try {
+            conn = getConnection();
+            conn.setAutoCommit(false);
             stmt = conn.prepareStatement("INSERT INTO CHECKLIST(" +
                     "  COD_UNIDADE, " +
                     "  COD_CHECKLIST_MODELO, " +
@@ -101,57 +70,106 @@ public class ChecklistDaoImpl extends DatabaseConnection implements ChecklistDao
             if (rSet.next()) {
                 checklist.setCodigo(rSet.getLong("CODIGO"));
                 final Long codUnidade = rSet.getLong("COD_UNIDADE");
-                insertRespostas(checklist, conn);
-                if (deveAbrirOs) {
-                    Injection
-                            .provideOrdemServicoDao()
-                            .processaChecklistRealizado(conn, codUnidade, checklist);
-                }
+                insertRespostas(conn, codUnidade, checklist);
+                Injection
+                        .provideOrdemServicoDao()
+                        .processaChecklistRealizado(conn, codUnidade, checklist);
                 veiculoDao.updateKmByPlaca(checklist.getPlacaVeiculo(), checklist.getKmAtualVeiculo(), conn);
+                conn.commit();
                 return checklist.getCodigo();
             } else {
                 throw new SQLException("Erro ao inserir o checklist");
             }
+        } catch (final Throwable t) {
+            if (conn != null) {
+                conn.rollback();
+            }
+
+            // Como esse método ainda não está refatorado para retornar um Throwable, encapsulamos o retorno em uma
+            // SQLException.
+            throw new SQLException(t);
         } finally {
-            close(stmt, rSet);
+            close(conn, stmt, rSet);
         }
     }
 
+    @NotNull
     @Override
-    public Checklist getByCod(Long codChecklist, String userToken) throws SQLException {
+    public Checklist getByCod(@NotNull final Long codChecklist, @NotNull final String userToken) throws SQLException {
         Connection conn = null;
         PreparedStatement stmt = null;
         ResultSet rSet = null;
         try {
             conn = getConnection();
-            stmt = conn.prepareStatement("SELECT " +
-                    "  C.CODIGO, " +
-                    "  C.COD_CHECKLIST_MODELO, " +
-                    "  C.DATA_HORA AT TIME ZONE ? AS DATA_HORA, " +
-                    "  C.DATA_HORA_IMPORTADO_PROLOG AT TIME ZONE ? AS DATA_HORA_IMPORTADO_PROLOG, " +
-                    "  C.KM_VEICULO, " +
-                    "  C.TEMPO_REALIZACAO, " +
-                    "  C.CPF_COLABORADOR, " +
-                    "  C.PLACA_VEICULO, " +
-                    "  C.TIPO, CO.NOME " +
-                    "FROM CHECKLIST C " +
-                    "  JOIN COLABORADOR CO " +
-                    "    ON CO.CPF = C.CPF_COLABORADOR " +
-                    "WHERE C.CODIGO = ?;");
-            final String id = TimeZoneManager.getZoneIdForToken(userToken, conn).getId();
-            stmt.setString(1, id);
-            stmt.setString(2, id);
-            stmt.setLong(3, codChecklist);
+            stmt = conn.prepareStatement("SELECT * FROM FUNC_CHECKLIST_GET_BY_CODIGO(F_COD_CHECKLIST := ?);");
+            stmt.setLong(1, codChecklist);
             rSet = stmt.executeQuery();
-            if (rSet.next()) {
-                final Checklist checklist = ChecklistConverter.createChecklist(rSet, false);
-                checklist.setListRespostas(getPerguntasRespostas(checklist));
-                return checklist;
-            } else {
-                throw new IllegalStateException("Checklist com o código: " + codChecklist + " não encontrado!");
+            PerguntaRespostaChecklist pergunta = null;
+            Long codChecklistAntigo = null, codChecklistAtual;
+            Long codPerguntaAntigo = null, codPerguntaAtual;
+            Checklist checklist = null;
+            boolean isFirstLine = true;
+            while (rSet.next()) {
+                codChecklistAtual = rSet.getLong("COD_CHECKLIST");
+                if (codChecklistAntigo == null) {
+                    codChecklistAntigo = codChecklistAtual;
+                }
+
+                codPerguntaAtual = rSet.getLong("COD_PERGUNTA");
+                if (codPerguntaAntigo == null) {
+                    codPerguntaAntigo = codPerguntaAtual;
+                }
+
+                if (isFirstLine) {
+                    checklist = ChecklistConverter.createChecklist(rSet, false);
+                    pergunta = ChecklistConverter.createPergunta(rSet);
+                    pergunta.setAlternativasResposta(new ArrayList<>());
+                    checklist.setListRespostas(new ArrayList<>());
+                    checklist.getListRespostas().add(pergunta);
+                    isFirstLine = false;
+                }
+
+                if (codChecklistAntigo.equals(codChecklistAtual)) {
+                    if (codPerguntaAntigo.equals(codPerguntaAtual)) {
+                        // Cria mais uma alternativa na pergunta atual.
+                        pergunta.getAlternativasResposta().add(ChecklistConverter.createAlternativaComResposta(rSet));
+                    } else {
+                        // Cria nova pergunta.
+                        pergunta = ChecklistConverter.createPergunta(rSet);
+                        pergunta.setAlternativasResposta(new ArrayList<>());
+                        checklist.getListRespostas().add(pergunta);
+
+                        // Cria primeira alternativa da nova pergunta.
+                        pergunta.getAlternativasResposta().add(ChecklistConverter.createAlternativaComResposta(rSet));
+                    }
+                } else {
+                    throw new IllegalStateException(
+                            "Esse método só está preparado para lidar com o retorno de um único checklist!");
+                }
+                codChecklistAntigo = codChecklistAtual;
+                codPerguntaAntigo = codPerguntaAtual;
             }
+
+            if (checklist == null) {
+                throw new IllegalStateException("Nenhum checklist encontrado com o código: " + codChecklist);
+            }
+
+            // Agora que já acabamos de criar, podemos forçar a contagem de itens OK/NOK a acontecer.
+            checklist.calculaQtdOkOrNok();
+
+            // Como a busca é feita ordenando pelo código, antes de retornar para o front nós ordenamos pela ordem de
+            // exibição das perguntas. Ignoramos a ordem de exibição das alternativas, não vale o overhead pelo que se
+            // ganha, atualmente, em exibição no front.
+            // O motivo de ordenarmos a busca pelo código ao invés de já direto pela ordem de exibição, é que atualmente
+            // a tabela de perguntas e alternativas não possuem nenhuma constraint que impeça a ordem de exibição de se
+            // repetir.
+            checklist
+                    .getListRespostas()
+                    .sort(Comparator.comparing(PerguntaRespostaChecklist::getOrdemExibicao));
+
+            return checklist;
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
     }
 
@@ -189,7 +207,7 @@ public class ChecklistDaoImpl extends DatabaseConnection implements ChecklistDao
             }
             return checklists;
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
     }
 
@@ -220,31 +238,72 @@ public class ChecklistDaoImpl extends DatabaseConnection implements ChecklistDao
             }
             return checklists;
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
     }
 
+    @NotNull
     @Override
-    public NovoChecklistHolder getNovoChecklistHolder(Long codUnidade, Long codModelo, String placa, char
-            tipoChecklis) throws SQLException {
-        final NovoChecklistHolder holder = new NovoChecklistHolder();
-        final ChecklistModeloDao checklistModeloDaoImpl = Injection.provideChecklistModeloDao();
-        final VeiculoDao veiculoDao = Injection.provideVeiculoDao();
-        holder.setCodigoModeloChecklist(codModelo);
-        holder.setListPerguntas(checklistModeloDaoImpl.getPerguntas(codUnidade, codModelo));
-        holder.setVeiculo(veiculoDao.getVeiculoByPlaca(placa, false));
-        return holder;
-    }
-
-    @Override
-    public Map<ModeloChecklist, List<String>> getSelecaoModeloChecklistPlacaVeiculo(Long codUnidade, Long codFuncao)
-            throws SQLException {
+    public FiltroRegionalUnidadeChecklist getRegionaisUnidadesSelecao(@NotNull final Long codColaborador)
+            throws Throwable {
         Connection conn = null;
         PreparedStatement stmt = null;
         ResultSet rSet = null;
-        final Map<ModeloChecklist, List<String>> modeloPlaca = new LinkedHashMap<>();
-        ModeloChecklist modelo = null;
-        List<String> placas = new ArrayList<>();
+        try {
+            conn = getConnection();
+            stmt = conn.prepareStatement(
+                    "SELECT * FROM FUNC_CHECKLIST_GET_REGIONAIS_UNIDADES_SELECAO(F_COD_COLABORADOR := ?);");
+            stmt.setLong(1, codColaborador);
+            rSet = stmt.executeQuery();
+            FiltroRegionalUnidadeChecklist filtro = null;
+            RegionalSelecaoChecklist regional = null;
+            Long codRegionalAntiga = null, codRegionalAtual;
+            boolean isFirstLine = true;
+            while (rSet.next()) {
+                codRegionalAtual = rSet.getLong("CODIGO_REGIONAL");
+                if (codRegionalAntiga == null) {
+                    codRegionalAntiga = codRegionalAtual;
+                }
+
+                if (isFirstLine) {
+                    filtro = new FiltroRegionalUnidadeChecklist(
+                            codColaborador,
+                            new ArrayList<>(),
+                            !rSet.getBoolean("REALIZACAO_CHECKLIST_DIFERENTES_UNIDADES_BLOQUEADO_EMPRESA"));
+                    regional = ChecklistConverter.createRegionalSelecao(rSet, new ArrayList<>());
+                    filtro.getRegionaisSelecao().add(regional);
+                    isFirstLine = false;
+                }
+
+                if (codRegionalAntiga.equals(codRegionalAtual)) {
+                    regional.getUnidadesVinculadas().add(ChecklistConverter.createUnidadeSelecao(rSet));
+                } else {
+                    regional = ChecklistConverter.createRegionalSelecao(rSet, new ArrayList<>());
+                    regional.getUnidadesVinculadas().add(ChecklistConverter.createUnidadeSelecao(rSet));
+                    filtro.getRegionaisSelecao().add(regional);
+                }
+                codRegionalAntiga = codRegionalAtual;
+            }
+
+            if (filtro == null) {
+                throw new IllegalStateException(
+                        "Dados de filtro não encontrados para o colaborador: " + codColaborador);
+            }
+
+            return filtro;
+        } finally {
+            close(conn, stmt, rSet);
+        }
+    }
+
+    @NotNull
+    @Override
+    public Map<ModeloChecklist, List<String>> getSelecaoModeloChecklistPlacaVeiculo(
+            @NotNull final Long codUnidade,
+            @NotNull final Long codFuncao) throws SQLException {
+        Connection conn = null;
+        PreparedStatement stmt = null;
+        ResultSet rSet = null;
         try {
             conn = getConnection();
             stmt = conn.prepareStatement("SELECT " +
@@ -271,6 +330,9 @@ public class ChecklistDaoImpl extends DatabaseConnection implements ChecklistDao
             stmt.setLong(1, codUnidade);
             stmt.setLong(2, codFuncao);
             rSet = stmt.executeQuery();
+            final Map<ModeloChecklist, List<String>> modeloPlaca = new LinkedHashMap<>();
+            ModeloChecklist modelo = null;
+            List<String> placas = new ArrayList<>();
             while (rSet.next()) {
                 // Primeira linha do Rset, cria o modelo, add a primeira placa.
                 if (modelo == null) {
@@ -296,12 +358,25 @@ public class ChecklistDaoImpl extends DatabaseConnection implements ChecklistDao
             if (modelo != null) {
                 modeloPlaca.put(modelo, placas);
             }
+            return modeloPlaca;
         } finally {
             close(conn, stmt, rSet);
         }
-        return modeloPlaca;
     }
 
+    @Override
+    public NovoChecklistHolder getNovoChecklistHolder(Long codUnidade, Long codModelo, String placa, char
+            tipoChecklis) throws SQLException {
+        final NovoChecklistHolder holder = new NovoChecklistHolder();
+        final ChecklistModeloDao checklistModeloDaoImpl = Injection.provideChecklistModeloDao();
+        final VeiculoDao veiculoDao = Injection.provideVeiculoDao();
+        holder.setCodigoModeloChecklist(codModelo);
+        holder.setListPerguntas(checklistModeloDaoImpl.getPerguntas(codUnidade, codModelo));
+        holder.setVeiculo(veiculoDao.getVeiculoByPlaca(placa, false));
+        return holder;
+    }
+
+    @NotNull
     @Override
     public DeprecatedFarolChecklist getFarolChecklist(@NotNull final Long codUnidade,
                                                       @NotNull final LocalDate dataInicial,
@@ -322,42 +397,42 @@ public class ChecklistDaoImpl extends DatabaseConnection implements ChecklistDao
             rSet = stmt.executeQuery();
             return ChecklistConverter.createFarolChecklist(rSet);
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
     }
 
-    @NotNull
     @Override
-    public Map<Long, AlternativaChecklistStatus> getItensStatus(@NotNull final Connection conn,
-                                                                @NotNull final Long codModelo,
-                                                                @NotNull final String placaVeiculo) throws Throwable {
+    public boolean getChecklistDiferentesUnidadesAtivoEmpresa(@NotNull final Long codEmpresa) throws Throwable {
+        Connection conn = null;
         PreparedStatement stmt = null;
         ResultSet rSet = null;
         try {
-            stmt = conn.prepareStatement("SELECT * FROM FUNC_CHECKLIST_OS_ALTERNATIVAS_ABERTURA_OS(?, ?);");
-            stmt.setLong(1, codModelo);
-            stmt.setString(2, placaVeiculo);
+            conn = getConnection();
+            stmt = conn.prepareStatement("SELECT * FROM " +
+                    "FUNC_CHECKLIST_REALIZACAO_DIFERENTES_UNIDADES_EMPRESA_BLOQUEADA(F_COD_EMPRESA := ?);");
+            stmt.setLong(1, codEmpresa);
             rSet = stmt.executeQuery();
-            final Map<Long, AlternativaChecklistStatus> alternativas = new HashMap<>();
-            while (rSet.next()) {
-                alternativas.put(
-                        rSet.getLong("COD_ALTERNATIVA"),
-                        ChecklistConverter.createAlternativaChecklistStatus(rSet));
+            if (rSet.next()) {
+                return !rSet.getBoolean("REALIZACAO_CHECKLIST_DIFERENTES_UNIDADES_BLOQUEADO_EMPRESA");
+            } else {
+                throw new SQLException("Erro ao verificar se a empresa está bloqueada para realizar checklist de " +
+                        "diferentes unidades");
             }
-            return alternativas;
         } finally {
-            close(stmt, rSet);
+            close(conn, stmt, rSet);
         }
     }
 
-    private void insertRespostas(Checklist checklist, Connection conn) throws SQLException {
+    private void insertRespostas(@NotNull final Connection conn,
+                                 @NotNull final Long codUnidade,
+                                 @NotNull final Checklist checklist) throws SQLException {
         PreparedStatement stmt = null;
         try {
             stmt = conn.prepareStatement("INSERT INTO CHECKLIST_RESPOSTAS "
                     + "(COD_UNIDADE, COD_CHECKLIST_MODELO, COD_CHECKLIST, COD_PERGUNTA, COD_ALTERNATIVA, RESPOSTA) "
-                    + "VALUES ((SELECT V.COD_UNIDADE FROM VEICULO V WHERE V.PLACA=?), ?, ?, ?, ?, ?)");
+                    + "VALUES (?, ?, ?, ?, ?, ?)");
             for (PerguntaRespostaChecklist resposta : checklist.getListRespostas()) {
-                stmt.setString(1, checklist.getPlacaVeiculo());
+                stmt.setLong(1, codUnidade);
                 stmt.setLong(2, checklist.getCodModelo());
                 stmt.setLong(3, checklist.getCodigo());
                 stmt.setLong(4, resposta.getCodigo());
@@ -398,17 +473,17 @@ public class ChecklistDaoImpl extends DatabaseConnection implements ChecklistDao
         try {
             conn = getConnection();
             stmt = conn.prepareStatement("SELECT " +
-                            "  CP.CODIGO AS COD_PERGUNTA, " +
-                            "  CP.ORDEM AS ORDEM_PERGUNTA, " +
-                            "  CP.PERGUNTA, " +
-                            "  CP.SINGLE_CHOICE, " +
-                            "  CAP.CODIGO AS COD_ALTERNATIVA, " +
-                            "  CAP.PRIORIDADE, " +
-                            "  CAP.ORDEM, " +
-                            "  CGI.COD_IMAGEM, " +
-                            "  CGI.URL_IMAGEM, " +
-                            "  CAP.ALTERNATIVA, " +
-                            "  CR.RESPOSTA " +
+                            "CP.CODIGO              AS COD_PERGUNTA," +
+                            "CP.ORDEM               AS ORDEM_PERGUNTA," +
+                            "CP.PERGUNTA            AS DESCRICAO_PERGUNTA," +
+                            "CP.SINGLE_CHOICE       AS PERGUNTA_SINGLE_CHOICE," +
+                            "CAP.CODIGO             AS COD_ALTERNATIVA," +
+                            "CAP.PRIORIDADE :: TEXT AS PRIORIDADE_ALTERNATIVA," +
+                            "CAP.ORDEM              AS ORDEM_ALTERNATIVA," +
+                            "CAP.ALTERNATIVA        AS DESCRICAO_ALTERNATIVA," +
+                            "CGI.COD_IMAGEM         AS COD_IMAGEM," +
+                            "CGI.URL_IMAGEM         AS URL_IMAGEM," +
+                            "CR.RESPOSTA            AS RESPOSTA " +
                             "FROM CHECKLIST C " +
                             "  JOIN CHECKLIST_RESPOSTAS CR " +
                             "    ON C.CODIGO = CR.COD_CHECKLIST " +
@@ -416,12 +491,10 @@ public class ChecklistDaoImpl extends DatabaseConnection implements ChecklistDao
                             "       AND C.COD_UNIDADE = CR.COD_UNIDADE " +
                             "  JOIN CHECKLIST_PERGUNTAS CP " +
                             "    ON CP.CODIGO = CR.COD_PERGUNTA " +
-                            "       AND CP.COD_UNIDADE = CR.COD_UNIDADE " +
                             "       AND CP.COD_CHECKLIST_MODELO = CR.COD_CHECKLIST_MODELO " +
                             "       AND CP.CODIGO = CR.COD_PERGUNTA " +
                             "  JOIN CHECKLIST_ALTERNATIVA_PERGUNTA CAP " +
                             "    ON CAP.CODIGO = CR.COD_ALTERNATIVA " +
-                            "       AND CAP.COD_UNIDADE = CR.COD_UNIDADE " +
                             "       AND CAP.COD_CHECKLIST_MODELO = CR.COD_CHECKLIST_MODELO " +
                             "       AND CAP.COD_PERGUNTA = CR.COD_PERGUNTA " +
                             "  LEFT JOIN CHECKLIST_GALERIA_IMAGENS CGI " +
@@ -435,7 +508,7 @@ public class ChecklistDaoImpl extends DatabaseConnection implements ChecklistDao
             rSet = stmt.executeQuery();
             return ChecklistConverter.createPerguntasRespostasChecklist(rSet);
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
     }